requirements.txt
discord.py
selenium
webdriver-manager
d20
psycopg
click
pytest
sqlalchemy